requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
django-decouple==2.1
django-unfold==0.65.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
